requirements.txt
python-dotenv==1.0.0
pandas==1.5.3
streamlit==1.19.0
plotly==5.13.1
requests==2.28.2
openpyxl==3.1.1
